requirements.txt
# ==========================================================
# requirements.txt – Dépendances Python pour Marais'R'Sense
# Version recommandée de Python : >=3.10
# ==========================================================

# --- Bibliothèques principales pour l'interface graphique ---
kivy>=2.3.0

# --- Communication MQTT ---
paho-mqtt>=2.0.0

# --- Gestion du matériel Unihiker et capteurs ---
pinpong>=1.2.0
unihiker>=1.0.0

# --- Capteurs de particules fines (optionnel) ---
sds011>=0.2.11

# --- Utilitaires réseau et série ---
getmac>=0.9.4
pyserial>=3.5

# --- Modules standards (inclus dans Python, pas besoin d'installation) ---
# ssl
# time

# ----------------------------------------------------------
# Pour installer toutes les dépendances :
# pip install -r requirements.txt
# ----------------------------------------------------------

